# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CampusFlow API
  version: 1.0.0
  description: >
    Stub generated from Zod schemas + route files (auth.ts, index.ts health/colleges,
    hackathons.ts). Not full coverage — P0 domains only. Generate full spec from Zod
    before public pilot.
servers:
  - url: http://localhost:4000
    description: Local dev
  - url: https://campusflow-api.onrender.com
    description: Render prod (placeholder)
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
...
paths:
...
  /api/colleges/register:
    post:
      summary: Register college (public, rate-limited 5/h)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, code, adminEmail]
              properties:
                name:
                  type: string
                code:
                  type: string
                address:
                  type: string
                phone:
                  type: string
                website:
                  type: string
                adminEmail:
                  type: string
                  format: email
      responses:
        '201':
          description: PENDING college created
        '400':
          description: Missing / duplicate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/colleges/{id}/departments:
    get:
      summary: Departments for a college (public)
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: id+name list
        '400':
          description: Invalid id (planned UUID 400)
...